requirements.txt
# Project requirements

# Core dependencies
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
pandas>=1.2.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
Pillow>=8.0.0
opencv-python>=4.5.0
tqdm>=4.50.0

# Data handling
kagglehub>=0.1.0

# Optional - for CUDA support
# torch>=1.8.0+cu111
# torchvision>=0.9.0+cu111